Finish splitting the basic tensor tests from the Enzyme differentiable functions ones
